io_uring: keep one copy of the write-lifecycle test helpers
Motivation:

Every test that checks a submitted buffer is released exactly once carried its own
awaitRefCntZero, four byte-identical copies differing only in whether the second
parameter was a ByteBuf or a DefaultFileRegion. The two tracker tests each carried
their own runOnEventLoop and BookkeepingTask as well.

Modification:

awaitRefCntZero takes a ReferenceCounted, which is all either caller needed, and
moves to IoUringRefCntZeroAwaiter. The tracker tests take runOnEventLoop and
BookkeepingTask from a shared AbstractWriteOperationTrackerTest. The two stay apart
because leak verification and the tracker harness have nothing to do with each other.

Result:

One definition of each. No test behaviour changes.
