+++ transport-native-io_uring/src/test/java/io/netty/channel/uring/IoUringSocketSendSzSendmsgZcTest.java
@@ -38,6 +38,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
+import static io.netty.channel.uring.IoUringRefCntZeroAwaiter.awaitRefCntZero;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -226,21 +227,5 @@ public class IoUringSocketSendSzSendmsgZcTest extends AbstractClientSocketTest {
                 }
             }
         }
-    }
-
-    private static boolean awaitRefCntZero(Channel channel, ByteBuf buffer, long timeout, TimeUnit unit)
-            throws InterruptedException {
-        CountDownLatch released = new CountDownLatch(1);
-        channel.eventLoop().execute(new Runnable() {
-            @Override
-            public void run() {
-                if (buffer.refCnt() == 0) {
-                    released.countDown();
-                } else {
-                    channel.eventLoop().schedule(this, 10, TimeUnit.MILLISECONDS);
-                }
-            }
-        });
-        return released.await(timeout, unit);
     }
 }
